requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
langchain-groq
faiss-cpu
pymupdf
python-dotenv
numpy
pandas
sentence-transformers
streamlit-shadcn-ui
